Split tests for each service. Add test target for ObjC.
